Create Experimental Two Axis Throttle
This version of the script has a modified logic throttle intended to convert the existing throttle calculation designed for a combined brake and throttle axis centered around 0 and map it onto seperate throttle (y axis) and brake (z axis) that run from negative 2^32 to positive 2^32 instead. I've also added the new axis to the diagnostic section.